+++ amd/src/mod_form.js
@@ -1,32 +1,23 @@
 define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_server_config', 'core/ajax'],
     function() {
-        var addServerButton = document.getElementById("id_add_server_button");
-        var missingConfig = document.getElementsByName("mumie_missing_config")[0];
-        var lmsSelectorUrl;
-        var systemLanguage;
-        var serverController = (function() {
-            var serverStructure;
-            var serverDropDown = document.getElementById("id_server");
+        const addServerButton = document.getElementById("id_add_server_button");
+        const missingConfig = document.getElementsByName("mumie_missing_config")[0];
+        let lmsSelectorUrl;
+        let systemLanguage;
+        const serverController = (function() {
+            let serverStructure;
+            const serverDropDown = document.getElementById("id_server");
 
             return {
                 init: function(structure) {
                     serverStructure = structure;
                     serverDropDown.onchange = function() {
                         courseController.updateOptions();
-                        langController.updateOptions();
-                        taskController.updateCompleteCourseVisibility();
                     };
                 },
                 getSelectedServer: function() {
-                    var selectedServerName = serverDropDown.options[serverDropDown.selectedIndex].text;
-
-                    for (var i in serverStructure) {
-                        var server = serverStructure[i];
-                        if (server.name == selectedServerName) {
-                            return server;
-                        }
-                    }
-                    return null;
+                    const selectedServerName = serverDropDown.options[serverDropDown.selectedIndex].text;
+                    return serverStructure.find(server => server.name === selectedServerName);
                 },
                 disable: function() {
                     serverDropDown.disabled = true;
@@ -38,10 +29,10 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
             };
         })();
 
-        var problemSelectorController = (function() {
-            var problemSelectorButton = document.getElementById('id_prb_selector_btn');
-            var problemSelectorWindow;
-            var mumieOrg = document.getElementsByName('mumie_org')[0].value;
+        const problemSelectorController = (function() {
+            const problemSelectorButton = document.getElementById('id_prb_selector_btn');
+            let problemSelectorWindow;
+            const mumieOrg = document.getElementsByName('mumie_org')[0].value;
 
             /**
              * Send a message to the problem selector window.
@@ -83,13 +74,15 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
              */
             function addMessageListener() {
                 window.addEventListener('message', (event) => {
-                    if (event.origin != lmsSelectorUrl) {
+                    if (event.origin !== lmsSelectorUrl) {
                         return;
                     }
-                    var importObj = JSON.parse(event.data);
+                    const importObj = JSON.parse(event.data);
+                    const isGraded = importObj.isGraded !== false;
                     try {
                         langController.setLanguage(importObj.language);
                         taskController.setSelection(importObj.link + '?lang=' + importObj.language);
+                        taskController.setIsGraded(isGraded);
                         sendSuccess();
                         window.focus();
                         displayProblemSelectedMessage();
@@ -118,6 +111,7 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
 
             return {
                 init: function() {
+                    const gradingType = taskController.getGradingType();
                     problemSelectorButton.onclick = function() {
                         problemSelectorWindow = window.open(
                             lmsSelectorUrl
@@ -130,6 +124,7 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                                 + langController.getSelectedLanguage()
                                 + '&origin=' + encodeURIComponent(window.location.origin)
                                 + '&uiLang=' + systemLanguage
+                                + '&gradingType=' + gradingType
                             , '_blank'
                         );
                     };
@@ -150,27 +145,23 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
             };
         })();
 
-        var courseController = (function() {
-            var courseDropDown = document.getElementById("id_mumie_course");
-            var coursefileElem = document.getElementsByName("mumie_coursefile")[0];
+        const courseController = (function() {
+            const courseDropDown = document.getElementById("id_mumie_course");
+            const coursefileElem = document.getElementsByName("mumie_coursefile")[0];
 
             /**
              * Add a new option the the 'MUMIE Course' drop down menu
              * @param {Object} course
              */
             function addOptionForCourse(course) {
-                var optionCourse = document.createElement("option");
-                var selectedLanguage = langController.getSelectedLanguage();
-                var name;
-                // If the currently selected server is not available on the server, we need to select another one.
+                const optionCourse = document.createElement("option");
+                const selectedLanguage = langController.getSelectedLanguage();
+                let name;
+                // If the currently selected language is not available on the server, we need to select another one.
                 if (!course.languages.includes(selectedLanguage)) {
                     name = course.name[0];
                 } else {
-                    for (var i in course.name) {
-                        if (course.name[i].language == selectedLanguage) {
-                            name = course.name[i];
-                        }
-                    }
+                    name = course.name.find(n => n.language === selectedLanguage);
                 }
                 optionCourse.setAttribute("value", name.value);
                 optionCourse.text = name.value;
@@ -188,74 +179,39 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                 init: function(isEdit) {
                     courseDropDown.onchange = function() {
                         updateCoursefilePath();
-                        langController.updateOptions();
-                        taskController.updateCompleteCourseVisibility();
                     };
                     courseController.updateOptions(isEdit ? coursefileElem.value : false);
                 },
                 getSelectedCourse: function() {
-                    var selectedCourseName = courseDropDown.options[courseDropDown.selectedIndex].text;
-                    var courses = serverController.getSelectedServer().courses;
-                    for (var i in courses) {
-                        var course = courses[i];
-                        for (var j in course.name) {
-                            if (course.name[j].value == selectedCourseName) {
-                                return course;
-                            }
-                        }
-                    }
-                    return null;
+                    const selectedCourseName = courseDropDown.options[courseDropDown.selectedIndex].text;
+                    const courses = serverController.getSelectedServer().courses;
+
+                    return courses.find(course => {
+                        return course.name.some(name => name.value === selectedCourseName);
+                    });
                 },
                 disable: function() {
                     courseDropDown.disabled = true;
                     removeChildElems(courseDropDown);
                 },
                 updateOptions: function() {
-                    var selectedCourseFile = coursefileElem.value;
+                    const selectedCourseFile = coursefileElem.value;
                     removeChildElems(courseDropDown);
                     courseDropDown.selectedIndex = 0;
-                    var courses = serverController.getSelectedServer().courses;
-                    for (var i in courses) {
-                        var course = courses[i];
-                        addOptionForCourse(course);
-                        if (course.coursefile == selectedCourseFile) {
-                            courseDropDown.selectedIndex = courseDropDown.childElementCount - 1;
-                        }
-                    }
+                    serverController.getSelectedServer().courses
+                        .forEach(course => {
+                            addOptionForCourse(course);
+                            if (course.coursefile === selectedCourseFile) {
+                                courseDropDown.selectedIndex = courseDropDown.childElementCount - 1;
+                            }
+                        });
                     updateCoursefilePath();
                 }
             };
         })();
 
-        var langController = (function() {
-            var languageDropDown = document.getElementById("id_language_dropdown");
-            var languageElem = document.getElementById("id_language");
-
-            /**
-             * Add a new option to the language drop down menu
-             * @param {string} lang the language to add
-             */
-            function addLanguageOption(lang) {
-                var optionLang = document.createElement("option");
-                optionLang.setAttribute("value", lang);
-                optionLang.text = lang;
-                languageDropDown.append(optionLang);
-            }
-
-            /**
-             * Update the selection of the drop down language with the given lang
-             * @param {string} lang
-             */
-            function setDropDownLanguage(lang) {
-                for (var i in languageDropDown.options) {
-                    var option = languageDropDown.options[i];
-                    if (option.value == lang) {
-                        languageDropDown.selectedIndex = i;
-                        courseController.updateOptions();
-                        return;
-                    }
-                }
-            }
+        const langController = (function() {
+            const languageElem = document.getElementById("id_language");
 
             /**
              * Check if the given language exists in the currently selected course.
@@ -266,14 +222,6 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                 return courseController.getSelectedCourse().languages.includes(lang);
             }
             return {
-                init: function() {
-                    languageDropDown.onchange = function() {
-                        languageElem.value = languageDropDown.options[languageDropDown.selectedIndex].text;
-                        courseController.updateOptions();
-                        taskController.setSelection();
-                    };
-                    langController.updateOptions(langController.getSelectedLanguage());
-                },
                 getSelectedLanguage: function() {
                     return languageElem.value;
                 },
@@ -282,40 +230,26 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                         throw new Error("Selected language not available");
                     }
                     languageElem.value = lang;
-                    setDropDownLanguage(lang);
-                },
-                disable: function() {
-                    languageDropDown.disabled = true;
-                    removeChildElems(languageDropDown);
-                },
-                updateOptions: function() {
-                    var currentLang = langController.getSelectedLanguage();
-                    removeChildElems(languageDropDown);
-                    languageDropDown.selectedIndex = 0;
-                    var languages = courseController.getSelectedCourse().languages;
-                    for (var i in languages) {
-                        var lang = languages[i];
-                        addLanguageOption(lang);
-                    }
-                    setDropDownLanguage(currentLang);
                 }
             };
         })();
 
-        var taskController = (function() {
-            var taskSelectionInput = document.getElementsByName("taskurl")[0];
-            var nameElem = document.getElementById("id_name");
-            var taskDisplayElement = document.getElementById("id_task_display_element");
-            var useCompleteCourseElem = document.getElementById("id_mumie_complete_course");
+        const taskController = (function() {
+            const taskSelectionInput = document.getElementsByName("taskurl")[0];
+            const nameElem = document.getElementById("id_name");
+            const taskDisplayElement = document.getElementById("id_task_display_element");
+            const isGradedElem = document.getElementById('id_mumie_isgraded');
+
 
             /**
              * Update the activity's name in the input field
              */
             function updateName() {
+                const newHeadline = getHeadline(taskController.getSelectedTask());
                 if (!isCustomName()) {
-                    nameElem.value = getHeadline(taskController.getSelectedTask());
+                    nameElem.value = newHeadline;
                 }
-                taskDisplayElement.value = getHeadline(taskController.getSelectedTask());
+                taskDisplayElement.value = newHeadline;
             }
 
             /**
@@ -324,7 +258,7 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
              * @return {boolean} True, if there is no headline with that name in all tasks
              */
             function isCustomName() {
-                if (nameElem.value.length == 0) {
+                if (nameElem.value.length === 0) {
                     return false;
                 }
                 return !getAllHeadlines().includes(nameElem.value);
@@ -333,19 +267,15 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
             /**
              * Get the task's headline for the currently selected language
              * @param {Object} task
-             * @returns  {string} the headline
+             * @returns  {string|null} the headline
              */
             function getHeadline(task) {
                 if (!task) {
                     return null;
                 }
-                for (var i in task.headline) {
-                    var localHeadline = task.headline[i];
-                    if (localHeadline.language == langController.getSelectedLanguage()) {
-                        return localHeadline.name;
-                    }
-                }
-                return null;
+                const selectedLanguage = langController.getSelectedLanguage();
+                const headlineWrapper = task.headline.find(localHeadline => localHeadline.language === selectedLanguage);
+                return headlineWrapper ? headlineWrapper.name : null;
             }
 
             /**
@@ -354,18 +284,9 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
              * @return {Object} Array containing all available tasks
              */
             function getAllTasks() {
-                var tasks = [];
-                for (var i in serverController.getAllServers()) {
-                    var server = serverController.getAllServers()[i];
-                    for (var j in server.courses) {
-                        var course = server.courses[j];
-                        for (var m in course.tasks) {
-                            var task = course.tasks[m];
-                            tasks.push(task);
-                        }
-                    }
-                }
-                return tasks;
+                return serverController.getAllServers()
+                    .flatMap(server => server.courses)
+                    .flatMap(course => course.tasks);
             }
 
             /**
@@ -373,21 +294,9 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
              * @returns {Object} Array containing all headlines
              */
             function getAllHeadlines() {
-                var headlines = [];
-                var tasks = getAllTasks();
-                tasks.push(getPseudoTaskFromCourse(courseController.getSelectedCourse()));
-                for (var i in tasks) {
-                    var task = tasks[i];
-                    for (var n in task.headline) {
-                        headlines.push(task.headline[n].name);
-                    }
-                }
-                var course = courseController.getSelectedCourse();
-                for (var j in course.name) {
-                    var name = course.name[j];
-                    headlines.push(name.value);
-                }
-                return headlines;
+                return getAllTasks().flatMap(task => task.headline)
+                    .map(headline => headline.name)
+                    .concat(courseController.getSelectedCourse().name.map(n => n.value));
             }
 
             /**
@@ -408,6 +317,16 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                 return link + "?lang=" + langController.getSelectedLanguage();
             }
 
+            /**
+             * Form inputs related to grades should be disabled, if the MUMIE Task is not graded.
+             */
+            function updateGradeEditability() {
+                const disabled = isGradedElem.value === '0';
+                document.getElementById('id_points').disabled = disabled;
+                document.getElementById('id_gradepass').disabled = disabled;
+                document.getElementById('id_duedate_enabled').disabled = disabled;
+                document.getElementById('id_gradecat').disabled = disabled;
+            }
 
             /**
              * Get a task that links to a course's overview page
@@ -429,69 +348,51 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                 };
             }
 
-            /**
-             * Returns true, if the user has chosen to select the entire course instead of a single problem
-             * @returns {boolean}
-             */
-            function useCompleteCourse() {
-                return useCompleteCourseElem.checked;
-            }
-
             return {
                 init: function() {
-                    taskController.updateCompleteCourseVisibility();
                     updateName();
-                    useCompleteCourseElem.onchange = function() {
-                        if (useCompleteCourse()) {
-                            var localizedLink = getLocalizedLink(courseController.getSelectedCourse().link);
-                            taskController.setSelection(localizedLink);
-                        } else {
-                            taskController.setSelection(null);
-                        }
-                        // Circumvent moodle bug that ignores "disabled" if visibility has changed.
-                        taskDisplayElement.disabled = "1";
-                    };
                 },
                 getSelectedTask: function() {
-                    var selectedLink = taskSelectionInput.value;
-                    var course = courseController.getSelectedCourse();
-                    var tasks = course.tasks.slice();
-                    tasks.push(getPseudoTaskFromCourse(course));
-                    for (var i in tasks) {
-                        var task = tasks[i];
-                        if (selectedLink == getLocalizedLinkFromTask(task)) {
-                            return task;
-                        }
-                    }
-                    return null;
+                    const selectedLink = taskSelectionInput.value;
+                    const tasks = courseController.getSelectedCourse()
+                        .tasks
+                        .slice();
+                    tasks.push(getPseudoTaskFromCourse(courseController.getSelectedCourse()));
+                    return tasks
+                        .find(task => getLocalizedLinkFromTask(task) === selectedLink);
                 },
                 setSelection: function(newSelection) {
-                    if (!newSelection && useCompleteCourse()) {
-                        newSelection = getLocalizedLink(courseController.getSelectedCourse().link);
-                    }
                     taskSelectionInput.value = newSelection;
                     updateName();
                 },
-                useCompleteCourse: function() {
-                    return useCompleteCourse();
+                setIsGraded: function(isGraded) {
+                    if (isGraded === null) {
+                        isGradedElem.value = null;
+                    }
+                    isGradedElem.value = isGraded ? '1' : '0';
+                    updateGradeEditability();
                 },
-                updateCompleteCourseVisibility: function() {
-                    var visible = courseController.getSelectedCourse().link;
-                    useCompleteCourseElem.parentElement.parentElement.parentElement.style =
-                     visible ? "display:block" : "display:none";
+                getGradingType: function() {
+                    const isGraded = isGradedElem.value;
+                    if (isGraded === '1') {
+                        return 'graded';
+                    } else if (isGraded === '0') {
+                        return 'ungraded';
+                    }
+                    return 'all';
                 }
             };
         })();
 
 
-        var multiTaskEditController = (function() {
-            var propertySelectionInputs = document.getElementsByName("mumie_multi_edit_property");
-            var selectedTaskProperties = document.getElementsByName("mumie_selected_task_properties")[0];
-            var selectedTaskProp = [];
-            var taskSelectionInputs = document.getElementsByName("mumie_multi_edit_task");
-            var selectedTasks = document.getElementsByName("mumie_selected_tasks")[0];
-            var selectedTaskIds = [];
-            var sectionInputs = document.getElementsByName("mumie_multi_edit_section");
+        const multiTaskEditController = (function() {
+            const propertySelectionInputs = document.getElementsByName("mumie_multi_edit_property");
+            const selectedTaskProperties = document.getElementsByName("mumie_selected_task_properties")[0];
+            let selectedTaskProp = [];
+            const taskSelectionInputs = document.getElementsByName("mumie_multi_edit_task");
+            const selectedTasks = document.getElementsByName("mumie_selected_tasks")[0];
+            let selectedTaskIds = [];
+            const sectionInputs = document.getElementsByName("mumie_multi_edit_section");
 
             /**
              * Push an element to an array, if it's not already included.
@@ -512,7 +413,7 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                 taskSelectionInputs.forEach(function(checkbox) {
                     checkbox.onchange = function() {
                         if (!checkbox.checked) {
-                            selectedTaskIds = selectedTaskIds.filter(elem => elem != checkbox.value);
+                            selectedTaskIds = selectedTaskIds.filter(elem => elem !== checkbox.value);
                         } else {
                             selectedTaskIds.push(checkbox.value);
                         }
@@ -528,7 +429,7 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                 propertySelectionInputs.forEach(function(checkbox) {
                     checkbox.onchange = function() {
                         if (!checkbox.checked) {
-                            selectedTaskProp = selectedTaskProp.filter(elem => elem != checkbox.value);
+                            selectedTaskProp = selectedTaskProp.filter(elem => elem !== checkbox.value);
                         } else {
                             selectedTaskProp.push(checkbox.value);
                         }
@@ -547,7 +448,7 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                             taskSelectionInputs.forEach(function(taskCheckbox) {
                                 if (taskCheckbox.getAttribute('section') === sectionCheckbox.value) {
                                     taskCheckbox.checked = false;
-                                    selectedTaskIds = selectedTaskIds.filter(elem => taskCheckbox.value != elem);
+                                    selectedTaskIds = selectedTaskIds.filter(elem => taskCheckbox.value !== elem);
                                 }
                             });
                         } else {
@@ -590,7 +491,6 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
             });
             serverController.disable();
             courseController.disable();
-            langController.disable();
             problemSelectorController.disable();
         }
 
@@ -598,8 +498,8 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
             init: function(contextid, prbSelectorUrl, lang) {
                 lmsSelectorUrl = prbSelectorUrl;
                 systemLanguage = lang;
-                var isEdit = document.getElementById("id_name").getAttribute('value');
-                var serverStructure = JSON.parse(document.getElementsByName('mumie_server_structure')[0].value);
+                const isEdit = document.getElementById("id_name").getAttribute('value');
+                const serverStructure = JSON.parse(document.getElementsByName('mumie_server_structure')[0].value);
                 if (isEdit && !serverConfigExists()) {
                     disableDropDownMenus('mumie_form_missing_server');
                 } else if (!serverStructure.length) {
@@ -608,7 +508,6 @@ define(['jquery', 'core/templates', 'core/modal_factory', 'auth_mumie/mumie_serv
                     serverController.init(serverStructure);
                     courseController.init(isEdit);
                     taskController.init(isEdit);
-                    langController.init();
                     multiTaskEditController.init();
                     problemSelectorController.init();
                 }